// src/app/enums/enums.ts
export enum NotificationTypes {
  Success = 'success',
  Error = 'error'
}

export enum UserTypes {
  Admin = 'Admin',
  Driver = 'Driver',
}

export enum UserTableFields {
  UserName = 'username',
  FirstName = 'firstname',
  LastName = 'lastname',
  Email = 'email',
  Type = 'type',
  Password = 'password',
  RepeatPassword = 'repeatPassword'
}

export enum UserAddFields {
  UserName = 'Username',
  FirstName = 'First name',
  LastName = 'Last name',
  Email = 'Email',
  Type = 'Type',
  Password = 'Password',
  RepeatPassword = 'Repeat Password'
}

export enum UserFormTypes {
  Text,
  Select
}